+++ components/Description/Description.module.scss
@@ -1,0 +1,15 @@
+.description {
+	line-height: 1.5;
+	font-size: 1.5rem;
+	text-align: center;
+}
+
+.code {
+	background: #fafafa;
+	cursor: pointer;
+	border-radius: 5px;
+	padding: 0.75rem;
+	font-size: 1.1rem;
+	font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
+		Bitstream Vera Sans Mono, Courier New, monospace;
+}
