+++ components/Grid/index.tsx
@@ -4,31 +4,33 @@ import Link from 'next/link';
 import React from 'react';
 import Card from '../Card';
 
+import styles from './Grid.module.scss';
+
 const Grid: React.FC = () => {
 	return (
-		<div className='gridLayout'>
-			<a href='https://www.twitter.com/Droid7ed' target="_blank">
+		<div className={styles.gridLayout}>
+			<a href='https://www.twitter.com/Droid7ed' target='_blank'>
 				<Card
 					title='Twitter'
 					icon={faTwitter}
-					textColor="beautifulBlue"
+					textColor='beautifulBlue'
 					shadowColor='beautifulBlueShadow'
 				/>
 			</a>
-			<a href="https://reddit.com/u/DroidZed" target="_blank">
-
+			<a href='https://reddit.com/u/DroidZed' target='_blank'>
 				<Card
 					title='u/DroidZed'
 					icon={faReddit}
-					textColor="bloodOrange"
+					textColor='bloodOrange'
 					shadowColor='bloodOrangeShadow'
 				/>
 			</a>
 			<Link href='/about'>
 				<a>
-					<Card title='About'
+					<Card
+						title='About'
 						icon={faInfoCircle}
-						textColor="goldenSun"
+						textColor='goldenSun'
 						shadowColor='goldenSunShadow'
 					/>
 				</a>
