+++ components/Title/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './Title.module.css';
+import styles from './Title.module.scss';
 
 const Title: React.FC = () => {
 	return (
