+++ package.json
@@ -15,7 +15,8 @@
     "@fortawesome/react-fontawesome": "^0.1.14",
     "next": "10.1.3",
     "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react-dom": "17.0.2",
+    "sass": "^1.32.12"
   },
   "devDependencies": {
     "@types/node": "^14.14.41",
