+++ pages/index.tsx
@@ -5,16 +5,20 @@ import HeadTag from '../components/HeadTag';
 import Title from '../components/Title';
 import { AppThemeContext } from '../store/themeStore';
 
+import styles from '../styles/Home.module.scss';
+
 function Home() {
 	const { theme } = useContext(AppThemeContext);
 
 	return (
 		<div className={!theme ? 'dark' : ''} id='darkBody'>
 			<div className='centered-flex top-container dark:top-container-dark'>
-				<HeadTag title='Home' />
-				<Title />
-				<Description />
-				<Grid />
+				<main className={styles.main}>
+					<HeadTag title='Home' />
+					<Title />
+					<Description />
+					<Grid />
+				</main>
 			</div>
 		</div>
 	);
