+++ styles/Home.module.scss
@@ -1,0 +1,8 @@
+.main {
+	padding: 5rem 0;
+	flex: 1;
+	display: flex;
+	flex-direction: column;
+	justify-content: center;
+	align-items: center;
+}
