--- styles/globals.css
@@ -1,104 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-@layer utilities {
-  @variants dark {
-    .top-container-dark {
-      background-color: #000;
-      color: #fff;
-    }
-  }
-
-  .beautifulBlueShadow:hover,
-  .beautifulBlueShadow:active,
-  .beautifulBlueShadow:focus {
-    box-shadow: rgba(6, 128, 220, 0.2) 0px 7px 29px 0px;
-  }
-
-  .bloodOrangeShadow:hover,
-  .bloodOrangeShadow:active,
-  .bloodOrangeShadow:focus {
-    box-shadow: rgba(255, 69, 0, 0.2) 0px 7px 29px 0px;
-  }
-
-  .goldenSunShadow:hover,
-  .goldenSunShadow:active,
-  .goldenSunShadow:focus {
-    box-shadow: rgba(255, 215, 0, 0.2) 0px 7px 29px 0px;
-  }
-
-  .beautifulBlue:active,
-  .beautifulBlue:focus,
-  .beautifulBlue:hover {
-    color: #0070f3;
-  }
-
-  .bloodOrange:hover,
-  .bloodOrange:focus,
-  .bloodOrange:active {
-    color: #ff4500;
-  }
-
-  .goldenSun:hover,
-  .goldenSun:focus,
-  .goldenSun:active {
-    color: #ffd700;
-  }
-
-  .centered-flex {
-    @apply flex;
-    @apply flex-col;
-    @apply items-center;
-    @apply justify-center;
-  }
-
-  .top-container {
-    @apply min-h-screen;
-    @apply py-0;
-    @apply px-2;
-  }
-
-  .main {
-    padding: 5rem 0;
-    flex: 1;
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-  }
-
-  .description {
-    line-height: 1.5;
-    font-size: 1.5rem;
-    text-align: center;
-  }
-
-  .code {
-    background: #fafafa;
-    cursor: pointer;
-    border-radius: 5px;
-    padding: 0.75rem;
-    font-size: 1.1rem;
-    font-family: Menlo, Monaco, Lucida Console, Liberation Mono,
-      DejaVu Sans Mono, Bitstream Vera Sans Mono, Courier New, monospace;
-  }
-
-  .gridLayout {
-    display: flex;
-    justify-content: center;
-    max-width: 800px;
-    margin-top: 3rem;
-  }
-
-  .logo {
-    height: 1em;
-  }
-
-  @media (max-width: 600px) {
-    .gridLayout {
-      width: 100%;
-      flex-direction: column;
-    }
-  }
-}
