+++ styles/globals.scss
@@ -1,0 +1,61 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@layer utilities {
+	@variants dark {
+		.top-container-dark {
+			background-color: #000;
+			color: #fff;
+		}
+	}
+
+	.beautifulBlueShadow:hover,
+	.beautifulBlueShadow:active,
+	.beautifulBlueShadow:focus {
+		box-shadow: rgba(6, 128, 220, 0.2) 0px 7px 29px 0px;
+	}
+
+	.bloodOrangeShadow:hover,
+	.bloodOrangeShadow:active,
+	.bloodOrangeShadow:focus {
+		box-shadow: rgba(255, 69, 0, 0.2) 0px 7px 29px 0px;
+	}
+
+	.goldenSunShadow:hover,
+	.goldenSunShadow:active,
+	.goldenSunShadow:focus {
+		box-shadow: rgba(255, 215, 0, 0.2) 0px 7px 29px 0px;
+	}
+
+	.beautifulBlue:active,
+	.beautifulBlue:focus,
+	.beautifulBlue:hover {
+		color: #0070f3;
+	}
+
+	.bloodOrange:hover,
+	.bloodOrange:focus,
+	.bloodOrange:active {
+		color: #ff4500;
+	}
+
+	.goldenSun:hover,
+	.goldenSun:focus,
+	.goldenSun:active {
+		color: #ffd700;
+	}
+
+	.centered-flex {
+		@apply flex;
+		@apply flex-col;
+		@apply items-center;
+		@apply justify-center;
+	}
+
+	.top-container {
+		@apply min-h-screen;
+		@apply py-0;
+		@apply px-2;
+	}
+}
